현:게시판;

+++ BNB/src/main/java/com/bit/bnb/hostboard/service/CommentService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.bit.bnb.hostboard.dao.CommentDao;
+import com.bit.bnb.hostboard.dao.HostBoardDao;
 import com.bit.bnb.hostboard.model.CommentVO;
 import com.bit.bnb.hostboard.model.ModifyCommentVO;
 
@@ -16,10 +17,19 @@ public class CommentService {
 	@Autowired
 	private CommentDao commentDao;
 	
+	@Autowired
+	private HostBoardDao hostBoardDao;
+	
 	//댓글작성
 	@Transactional
-	public int writeComment(CommentVO comment) {
-		return commentDao.insertComment(comment);
+	public int writeComment(CommentVO comment, int postNo) {
+		int resultCnt = 0;
+		int resultTmp = hostBoardDao.upCommentCnt(postNo);
+		
+		if(resultTmp == 1) {
+			resultCnt = commentDao.insertComment(comment);
+		}
+		return resultCnt;
 	}
 	
 	// 해당 포스트의 댓글 리스트 불러오기
@@ -30,8 +40,14 @@ public class CommentService {
 	
 	// 코멘트 삭제
 	@Transactional
-	public int deleteComment(int commentNo) {
-		return commentDao.deleteComment(commentNo);
+	public int deleteComment(int commentNo, int postNo) {
+		int resultCnt = 0; 
+		int resultTmp = hostBoardDao.downCommentCnt(postNo);
+
+		if(resultTmp == 1) {
+			resultCnt = commentDao.deleteComment(commentNo);
+		}
+		return resultCnt;
 	}
 	
 	// 댓글 한개 가져오기
@@ -45,6 +61,5 @@ public class CommentService {
 	public int modifyComment(ModifyCommentVO modifyComment) {
 		return commentDao.modifyComment(modifyComment); 
 	}
-	
 	
 }
